Refactor Photos component logic
Before I was using grid base layout for the photos, which - in a nutshell - dynamically calculates spans so that the images fit nicely. I found the method in one of the Stephen Grider's videos.

I replaced that with method similar to the one used by Unsplash on their website. It uses `column-count` css propery.

* **IMPORTANT** I've been struggling to prevent the images from splitting into two parts. The solution was to use `break-inside: avoid` on the image's parent div.